Refactored the tests
